fix: adding type definitions
