+++ src/v1/destinations/iterable/networkHandler.ts
@@ -3,12 +3,7 @@ import { processAxiosResponse } from '../../../adapters/utils/networkUtils';
 import { BULK_ENDPOINTS } from '../../../v0/destinations/iterable/config';
 import { GenericStrategy } from './strategies/generic';
 import { TrackIdentifyStrategy } from './strategies/track-identify';
-
-type ResponseParams = {
-  destinationRequest: {
-    endpoint: string;
-  };
-};
+import { GenericProxyHandlerInput } from './types';
 
 const strategyRegistry: { [key: string]: any } = {
   [TrackIdentifyStrategy.name]: new TrackIdentifyStrategy(),
@@ -22,7 +17,7 @@ const getResponseStrategy = (endpoint: string) => {
   return strategyRegistry[GenericStrategy.name];
 };
 
-const responseHandler = (responseParams: ResponseParams) => {
+const responseHandler = (responseParams: GenericProxyHandlerInput) => {
   const { destinationRequest } = responseParams;
   const strategy = getResponseStrategy(destinationRequest.endpoint);
   return strategy.handleResponse(responseParams);
