+++ src/v1/destinations/iterable/strategies/base.ts
@@ -2,28 +2,24 @@ import { TAG_NAMES } from '@rudderstack/integrations-lib';
 import { getDynamicErrorType } from '../../../../adapters/utils/networkUtils';
 import { isHttpStatusSuccess } from '../../../../v0/util';
 import { TransformerProxyError } from '../../../../v0/util/errorTypes';
-import { DestinationResponse, ResponseParams } from '../types';
+import { GenericProxyHandlerInput } from '../types';
 
 // Base strategy is the base class for all strategies in Iterable destination
 class BaseStrategy {
-  handleResponse(responseParams: { destinationResponse: DestinationResponse }): void {
+  handleResponse(responseParams: GenericProxyHandlerInput): void {
     const { destinationResponse } = responseParams;
     const { status } = destinationResponse;
 
     if (!isHttpStatusSuccess(status)) {
-      return this.handleError({
-        destinationResponse,
-        rudderJobMetadata: [],
-      });
+      return this.handleError(responseParams);
     }
 
     return this.handleSuccess(responseParams);
   }
 
-  handleError(responseParams: ResponseParams): void {
+  handleError(responseParams: GenericProxyHandlerInput): void {
     const { destinationResponse, rudderJobMetadata } = responseParams;
     const { response, status } = destinationResponse;
-    // @ts-expect-error: not sure if `response.message` is correct or needed
     const responseMessage = response.params || response.msg || response.message;
     const errorMessage = JSON.stringify(responseMessage) || 'unknown error format';
 
@@ -49,3 +45,11 @@ class BaseStrategy {
 }
 
 export { BaseStrategy };
+
+// TODO:
+/**
+ * 1) fix return types appropriately
+ * 2) use pre-declared types, rather than adding our own types
+ * 3) no need for unnecessary refactors
+ * 4) add different implementations for handle Errors
+ */
